+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.10.0
 * `NEW` workspace: supports `.dll`(`.so`) in `require`
-* `NEW` folding: `---@class` and docs of function
+* `NEW` folding: `---@class`, `--#region` and docs of function
 * `CHG` supports `~` in command line
 * `CHG` completion: improve workspace words
 * `CHG` completion: show words in string